tests/dnsd: read config from file
Make the <dns> tag in a test case control what is stored there. Also
documented. Make test 2102 use the new tag.

Lets the test case config the A and AAAA contents the server replies
with. If none are provided, a fixed internal default is used. For now
this only supports one address of each family.