An ugly fix to solve the reliability issue with the zero copy version
As explained in the previous commit, when curl is a little bit too slow, the connection was failing.
The reason was that stream_resp_read was called a second time by curl and quiceh_h3_recv_body_v3 returns the same packet if it wasn't consumed. That completely mess up the quiceh consume mechanism.
Now we enforce that when a packet is returned by quiceh, it has to be consumed before we can call quiceh_h3_recv_body_v3

To be clean, we need a new feature in quiceh to be able to get multiple packets without consuming them directly